Update loading wrapper file structure
